Rtor names of files, separate init of service and send email function
